+++ clicks/thermo21/lib/src/thermo21.c
@@ -1,0 +1,106 @@
+/****************************************************************************
+** Copyright (C) 2020 MikroElektronika d.o.o.
+** Contact: https://www.mikroe.com/contact
+**
+** Permission is hereby granted, free of charge, to any person obtaining a copy
+** of this software and associated documentation files (the "Software"), to deal
+** in the Software without restriction, including without limitation the rights
+** to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+** copies of the Software, and to permit persons to whom the Software is
+** furnished to do so, subject to the following conditions:
+** The above copyright notice and this permission notice shall be
+** included in all copies or substantial portions of the Software.
+**
+** THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND,
+** EXPRESS OR IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES
+** OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT.
+** IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM,
+** DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT
+** OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE
+**  USE OR OTHER DEALINGS IN THE SOFTWARE.
+****************************************************************************/
+
+/*!
+ * @file thermo21.c
+ * @brief Thermo 21 Click Driver.
+ */
+
+#include "thermo21.h"
+
+/**
+ * @brief Help macros.
+ * @details Definition of help macros.
+ */
+#define DUMMY       0x00
+#define RESOLUTION  32.0
+
+void thermo21_cfg_setup ( thermo21_cfg_t *cfg ) 
+{
+    cfg->sck  = HAL_PIN_NC;
+    cfg->miso = HAL_PIN_NC;
+    cfg->mosi = HAL_PIN_NC;
+    cfg->cs   = HAL_PIN_NC;
+
+    cfg->spi_speed   = 100000;
+    cfg->spi_mode    = SPI_MASTER_MODE_0;
+    cfg->cs_polarity = SPI_MASTER_CHIP_SELECT_POLARITY_ACTIVE_LOW;
+}
+
+err_t thermo21_init ( thermo21_t *ctx, thermo21_cfg_t *cfg ) 
+{
+    spi_master_config_t spi_cfg;
+
+    spi_master_configure_default( &spi_cfg );
+
+    spi_cfg.sck  = cfg->sck;
+    spi_cfg.miso = cfg->miso;
+    spi_cfg.mosi = cfg->mosi;
+
+    ctx->chip_select = cfg->cs;
+
+    if ( SPI_MASTER_ERROR == spi_master_open( &ctx->spi, &spi_cfg ) ) 
+    {
+        return SPI_MASTER_ERROR;
+    }
+
+    if ( SPI_MASTER_ERROR == spi_master_set_default_write_data( &ctx->spi, DUMMY ) ) 
+    {
+        return SPI_MASTER_ERROR;
+    }
+
+    if ( SPI_MASTER_ERROR == spi_master_set_mode( &ctx->spi, cfg->spi_mode ) ) 
+    {
+        return SPI_MASTER_ERROR;
+    }
+
+    if ( SPI_MASTER_ERROR == spi_master_set_speed( &ctx->spi, cfg->spi_speed ) ) 
+    {
+        return SPI_MASTER_ERROR;
+    }
+
+    spi_master_set_chip_select_polarity( cfg->cs_polarity );
+    spi_master_deselect_device( ctx->chip_select );
+
+    return SPI_MASTER_SUCCESS;
+}
+
+err_t thermo21_generic_read ( thermo21_t *ctx, uint8_t *data_out, uint8_t len ) 
+{
+    spi_master_select_device( ctx->chip_select );
+    err_t error_flag = spi_master_read( &ctx->spi, data_out, len );
+    spi_master_deselect_device( ctx->chip_select );
+
+    return error_flag;
+}
+
+err_t thermo21_temperature_read ( thermo21_t *ctx, float *data_out )
+{
+    uint8_t temp_buf[ 2 ] = { 0 };
+    err_t ret_val = thermo21_generic_read( ctx, temp_buf, 2 );
+    int16_t temp_data = ( ( int16_t )temp_buf[ 0 ] << 10 ) | ( ( int16_t )temp_buf[ 1 ] << 2 );
+    temp_data >>= 2;
+    *data_out = temp_data / RESOLUTION;
+    return ret_val;
+}
+
+// ------------------------------------------------------------------------- END